// src/app/core/models/tuttocitta-geocode-responce.interface.ts
export interface TuttocittaGeocodeResponceInterface {
    f: number;
    l: string;
    q: string;
    r: [{
        cap: string;
        code: number;
        com: string;
        lat: number;
        lon: number;
        pcode: number;
        prov: string;
        req: string;
        stato: string;
        t: number;
        topo: string;
    }];
    t: string;
}